third_party/WebKit/LayoutTests/inspector/console/console-uncaught-exception-in-eval-expected.txt

CONSOLE ERROR: line 3: Uncaught Error: Exception in eval:with sourceURL
CONSOLE ERROR: line 3: Uncaught Error: Exception in eval:anonymous
Tests that when uncaught exception in eval'ed script ending with //# sourceURL=url is logged into console, its stack trace will have the url as the script source. Bug 47252.

evalURL.js:3 Uncaught Error: Exception in eval:with sourceURL
    at b (evalURL.js:3)
    at a (evalURL.js:7)
    at evalSource (evalURL.js:10)
    at eval (evalURL.js:11)
    at eval (<anonymous>)
b @ evalURL.js:3
a @ evalURL.js:7
evalSource @ evalURL.js:10
(anonymous) @ evalURL.js:11
setTimeout (async)
doEvalWithSourceURL @ console-uncaught-exc…ion-in-eval.html:20
(anonymous) @ VM:1
VM:3 Uncaught Error: Exception in eval:anonymous
    at b (eval at <anonymous> (unknown source), <anonymous>:3:11)
    at a (eval at <anonymous> (unknown source), <anonymous>:7:5)
    at evalSource (eval at <anonymous> (unknown source), <anonymous>:10:3)
    at eval (eval at <anonymous> (unknown source), <anonymous>:11:3)
    at eval (<anonymous>)
b @ VM:3
a @ VM:7
evalSource @ VM:10
(anonymous) @ VM:11
setTimeout (async)
doAnonymousEvalWith @ console-uncaught-exc…ion-in-eval.html:25
(anonymous) @ VM:1
